h5_tf/requirements.txt
tensorflow
Pillow
numpy